+++ server/internal/capture/gst/gst.c
