+++ server/internal/capture/manager.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"time"
 
+	"m1k1o/neko/internal/capture/gst"
 	"m1k1o/neko/internal/desktop/xorg"
-	"m1k1o/neko/internal/gst"
 	"m1k1o/neko/internal/types"
 	"m1k1o/neko/internal/types/config"
